refactor(studio): use subfolder in shared directory
