+++ apps/studio/schemas/documents/news.category.ts
@@ -2,7 +2,7 @@ import { RiBookletLine } from 'react-icons/ri';
 import { defineType } from 'sanity';
 
 import { general } from '@/shared/field-groups';
-import { slugField, titleField } from '@/shared/general-fields';
+import { slugField, titleField } from '@/shared/fields/general';
 
 const newsCategory = defineType({
 	title: 'News-Kategorie',
