+++ apps/studio/schemas/single-pages/home/groups.ts
@@ -1,7 +1,7 @@
 import { RiLinksLine } from 'react-icons/ri';
 import { defineField } from 'sanity';
 
-import { subTitleField, titleField } from '@/shared/general-fields';
+import { subTitleField, titleField } from '@/shared/fields/general';
 
 const groupsField = defineField({
 	title: 'Gruppen',
