+++ apps/studio/schemas/single-pages/privacy.ts
@@ -1,10 +1,10 @@
 import { RiSettings5Line } from 'react-icons/ri';
 import { defineField, defineType } from 'sanity';
 
-import { addressField, emailField, phoneField } from '@/shared/contact-fields';
 import { contact, general, meta } from '@/shared/field-groups';
-import { defaultPageFields, getHiddenSlugField } from '@/shared/general-fields';
-import { metaField } from '@/shared/meta-fields';
+import { addressField, emailField, phoneField } from '@/shared/fields/contact';
+import { defaultPageFields, getHiddenSlugField } from '@/shared/fields/general';
+import { metaField } from '@/shared/fields/meta';
 import { getRequiredRole } from '@/shared/validation-rules';
 
 const privacyPage = defineType({
