+++ apps/studio/shared/fields/content.ts
@@ -1,6 +1,6 @@
 import { defineField } from 'sanity';
 
-import { getRequiredRole } from './validation-rules';
+import { getRequiredRole } from '../validation-rules';
 
 export const contactPersonsField = defineField({
 	title: 'Ansprechpartner',
